Add bitcode support
This commit adds support for bitcode Encode and Decode. It does so gated
behind a feature flag. The bitcode version is set to 0, with default
features turned off, only derive feature added, and optional set to
true. This allows libraries to specify the verison of bitcode they would
prefer, and avoids using any other functionality of bitcode.